Trait poa_rpc::PoARuntimeApi

pub trait PoARuntimeApi<Block, AccountId, Balance>: Core<Block> where
    Block: Block,
    AccountId: Codec + MaybeDisplay + MaybeFromStr,
    Balance: Codec + MaybeDisplay + MaybeFromStr, {
    pub fn get_treasury_account(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<AccountId, Self::Error> { ... }

    pub fn get_treasury_account_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<AccountId, Self::Error> { ... }

    pub fn get_treasury_balance(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Balance, Self::Error> { ... }

    pub fn get_treasury_balance_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Balance, Self::Error> { ... }

    pub fn get_total_emission_in_epoch(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        epoch_no: u32
    ) -> Result<Balance, Self::Error> { ... }

    pub fn get_total_emission_in_epoch_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        epoch_no: u32
    ) -> Result<Balance, Self::Error> { ... }
}

Provided methods

pub fn get_treasury_account(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<AccountId, Self::Error>

Return account address of treasury. The account address can then be used to query the chain for balance

pub fn get_treasury_account_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<AccountId, Self::Error>

Return account address of treasury. The account address can then be used to query the chain for balance

pub fn get_treasury_balance(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Balance, Self::Error>

Return free balance of treasury account. In the context of PoA, only free balance makes sense for treasury. But just in case, to check all kinds of balance (locked, reserved, etc), get the account address with above call and query the chain.

pub fn get_treasury_balance_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Balance, Self::Error>

Return free balance of treasury account. In the context of PoA, only free balance makes sense for treasury. But just in case, to check all kinds of balance (locked, reserved, etc), get the account address with above call and query the chain.

pub fn get_total_emission_in_epoch(
&self,
__runtime_api_at_param__: &BlockId<Block>,
epoch_no: u32
) -> Result<Balance, Self::Error>

Return total (validators + treasury) emission rewards for given epoch

pub fn get_total_emission_in_epoch_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
epoch_no: u32
) -> Result<Balance, Self::Error>

Return total (validators + treasury) emission rewards for given epoch

Loading content...