Trait poa_rpc::PoAApi

pub trait PoAApi<BlockHash, AccountId, Balance>: Sized + Send + Sync + 'static {
    fn treasury_account(&self, at: Option<BlockHash>) -> Result<AccountId>;

    fn treasury_balance(&self, at: Option<BlockHash>) -> Result<Balance>;

    fn get_total_emission_in_epoch(
        &self, 
        epoch_no: EpochNo, 
        at: Option<BlockHash>
    ) -> Result<Balance>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        AccountId: Send + Sync + 'static + Serialize,
        Balance: Send + Sync + 'static + Serialize,
    { ... }
}

Required methods

fn treasury_account(&self, at: Option<BlockHash>) -> Result<AccountId>

Return account address of treasury. The account address can then be used to query the chain for balance

fn treasury_balance(&self, at: Option<BlockHash>) -> Result<Balance>

Return free balance of treasury account. In the context of PoA, only free balance makes sense for treasury. But just in case, to check all kinds of balance (locked, reserved, etc), get the account address with above call and query the chain.

fn get_total_emission_in_epoch(
&self,
epoch_no: EpochNo,
at: Option<BlockHash>
) -> Result<Balance>

Return total (validators + treasury) emission rewards for given epoch

Loading content...
Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
AccountId: Send + Sync + 'static + Serialize,
Balance: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...